Flask/requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
python-socketio==5.9.0
python-engineio==4.8.0
gevent==23.9.1
gevent-websocket==0.10.1
Werkzeug==3.0.1
python-dotenv==1.0.0